Use create_table? instead of create_table!
If table has already been created, don't drop it and try to
recreate it, just skip it.  Dropping it no longer works now
that foreign key constraints are used, as it tries to drop
the vocabulary table before the tables that depend on
vocabulary.

While here, instead of using a separate primary_key method
call, just use the :primary_key=>true option for simple
primary keys.  The :null=>false can be dropped for such
primary keys, since PRIMARY KEY implies NOT NULL.

Also, foreign keys don't need to specify :key, since
the default is to reference the primary key of the related
table.