Refactor to allow loading of current IMEDS vocabulary data
There are a lot of interrelated changes here.  The reason behind
them is the IMEDS data format changed significantly, which made
the vocabulary data diverage in format from the cdm data, so
refactoring and adding some methods to the Loader subclasses
was necessary so that the same Loader superclass would work with
both.

Drop the headers hash, instead storing the headers in the all_files
data structure.  Change all_files from being a hash to being an
array, since it doesn't seem to be used as a hash.

In fast load, use symbols instead of strings, as apprantly adapter
is now a symbol.

Add postgres_copy_into_options and ruby_csv_options methods, which
the vocab loader overrides to deal with the nonstandard format
IMEDS uses for the vocabulary data.

Set the binary flag when reading files, to avoid encoding errors,
as the IMEDS data does not use UTF-8 format.

Use Sequel.qualify instead of constructing double underscored
symbols at runtime.

Set ASCII 242 as the field separator, and ASCII 255 as the quote
character.  The IMEDS data doesn't use a quote character, but
includes the quote character as a data character.  Filter out
carriage returns from the IMEDS data before splitting it, as both
PostgreSQL and ruby's CSV library don't allow carriage returns
in unquoted fields.